Remove "Storage" namespace
I didn't notice in the example we were following, but "Storage" was his
example of a namespace of a Gem. That's not really necessary here.
