+++ app/models/adapters.rb
@@ -1,0 +1,1 @@
+module Adapters; end
