+++ app/models/entities.rb
@@ -1,0 +1,1 @@
+module Entities; end
