+++ app/models/models.rb
@@ -1,0 +1,1 @@
+module Models; end
