+++ app/models/models/active_record.rb
@@ -1,0 +1,3 @@
+module Models
+  module ActiveRecord; end
+end
