+++ app/models/query_objects.rb
@@ -1,0 +1,1 @@
+module QueryObjects; end
