+++ app/models/query_objects/find_campus.rb
@@ -1,0 +1,15 @@
+require_relative 'query_object'
+
+module QueryObjects
+  class FindCampus
+    include QueryObject
+
+    def initialize(repository = Repositories::CampusRepository.new)
+      super
+    end
+
+    def call
+      repository.load_all
+    end
+  end
+end
