+++ app/models/query_objects/query_object.rb
@@ -1,0 +1,10 @@
+module QueryObjects
+  module QueryObject
+    def initialize(repository)
+      @repository = repository
+    end
+
+    private
+    attr_reader :repository
+  end
+end
