+++ app/models/repositories.rb
@@ -1,0 +1,1 @@
+module Repositories; end
