--- app/models/storage.rb
@@ -1,1 +1,0 @@
-module Storage; end
