--- app/models/storage/adapters.rb
@@ -1,3 +1,0 @@
-module Storage
-  module Adapters; end
-end
