--- app/models/storage/entities.rb
@@ -1,3 +1,0 @@
-module Storage
-  module Entities; end
-end
