--- app/models/storage/models.rb
@@ -1,3 +1,0 @@
-module Storage
-  module Models; end
-end
