--- app/models/storage/models/active_record.rb
@@ -1,5 +1,0 @@
-module Storage
-  module Models
-    module ActiveRecord; end
-  end
-end
