--- app/models/storage/models/active_record/campus.rb
@@ -1,14 +1,0 @@
-module Storage
-  module Models
-    module ActiveRecord
-      class Campus < ::ActiveRecord::Base
-
-        def self.orm_adapter
-          Storage::Adapters::ActiveRecord
-        end
-
-        self.primary_key = "abbreviation"
-      end
-    end
-  end
-end
