--- app/models/storage/query_objects.rb
@@ -1,3 +1,0 @@
-module Storage
-  module QueryObjects; end
-end
