--- app/models/storage/query_objects/query_object.rb
@@ -1,12 +1,0 @@
-module Storage
-  module QueryObjects
-    module QueryObject
-      def initialize(repository)
-        @repository = repository
-      end
-
-      private
-      attr_reader :repository
-    end
-  end
-end
