--- app/models/storage/repositories.rb
@@ -1,3 +1,0 @@
-module Storage
-  module Repositories; end
-end
